fix clear assets
