+++ src/Console/RebuildPlatformUi.php
@@ -3,7 +3,7 @@
 namespace Enjin\PlatformUi\Console;
 
 use Illuminate\Console\Command;
-use Illuminate\Support\Str;
+use Illuminate\Support\Facades\File;
 
 class RebuildPlatformUi extends Command
 {
@@ -14,6 +14,8 @@ class RebuildPlatformUi extends Command
 
     public function handle()
     {
+        $this->clearPlatformUIAssets();
+
         $this->comment('Installing Platform UI dependencies...');
         exec('npm install --prefix ' . $this->BASE_DIR . ' 2> /dev/null');
 
@@ -24,5 +26,14 @@ class RebuildPlatformUi extends Command
         $this->callSilent('vendor:publish', ['--tag' => 'platform-ui-assets', '--force' => true]);
 
         $this->info('Platform UI package Installed successfully.');
+    }
+
+    public function clearPlatformUIAssets()
+    {
+        $path = public_path('vendor/platform-ui');
+
+        if (File::exists($path)) {
+            File::deleteDirectory($path);
+        }
     }
 }
